+++ src/platform/win32/wglext.inl
@@ -676,7 +676,7 @@ void cleanup(void)
 bool checkExtension(const char* extension, bool use_dummy)
 {
 	if(!extension) return false;
-	const int ext_len = std::strlen(extension);
+	const int ext_len = (int)std::strlen(extension);
 	if(ext_len <= 0) return false;
 	if(use_dummy) 
 	{
